resume-ai-project/resume-ai-project/ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
python-dotenv==1.0.1
gunicorn==21.2.0
